#5 compiles on mac, thousands of warnings (#9)
* #5 compiles on mac, thousands of warnings

* #5 Finalized mac building and improved GLSL compilation

* Updated README and added CMake helper for Wasabi-based projects

+++ src/Wasabi/Animations/WSkeletalAnimation.cpp
@@ -246,7 +246,7 @@ WError WSkeleton::CreateKeyFrame(WBone* baseBone, float fTime) {
 
 	f->baseBone = new WBone();
 	uint32_t index = 0;
-	f->ConstructHierarchyFromBase(baseBone, f->baseBone, baseBone->GetIndex() == -1 ? &index : nullptr);
+	f->ConstructHierarchyFromBase(baseBone, f->baseBone, baseBone->GetIndex() == (uint32_t)-1 ? &index : nullptr);
 
 	WError err = WError(W_SUCCEEDED);
 
@@ -304,10 +304,10 @@ void WSkeleton::SetSubAnimationBaseBone(uint32_t subAnimation, uint32_t boneInde
 
 	if (boneIndex != -1) {
 		uint32_t indexInVector = (uint32_t)-1;
-		for (uint32_t i = 0; i < ((_WSkeletalFrame*)WAnimation::m_frames[0])->boneV.size() && indexInVector == -1; i++)
+		for (uint32_t i = 0; i < ((_WSkeletalFrame*)WAnimation::m_frames[0])->boneV.size() && indexInVector == (uint32_t)-1; i++)
 			if (((_WSkeletalFrame*)WAnimation::m_frames[0])->boneV[i]->GetIndex() == boneIndex)
 				indexInVector = i;
-		if (indexInVector == -1)
+		if (indexInVector == (uint32_t)-1)
 			return;
 
 		//get all involved bones, start with the parent (at boneIndex) and store all children indices
@@ -361,7 +361,7 @@ void WSkeleton::Update(float fDeltaTime) {
 				if (curSubAnim->parentSubAnimation != -1) {
 					uint32_t boneIndex = curSubAnim->boneIndices[0];
 					uint32_t boneInVector = (uint32_t)-1;
-					for (uint32_t k = 0; k < curFrame->boneV.size() && boneInVector == -1; k++)
+					for (uint32_t k = 0; k < curFrame->boneV.size() && boneInVector == (uint32_t)-1; k++)
 						if (curFrame->boneV[k]->GetIndex() == boneIndex)
 							boneInVector = k;
 					WBone* curFrameBone = curFrame->boneV[boneInVector];
@@ -382,7 +382,7 @@ void WSkeleton::Update(float fDeltaTime) {
 				for (uint32_t i = 0; i < curSubAnim->boneIndices.size(); i++) {
 					uint32_t boneIndex = curSubAnim->boneIndices[i];
 					uint32_t boneInVector = (uint32_t)-1;
-					for (uint32_t k = 0; k < curFrame->boneV.size() && boneInVector == -1; k++)
+					for (uint32_t k = 0; k < curFrame->boneV.size() && boneInVector == (uint32_t)-1; k++)
 						if (curFrame->boneV[k]->GetIndex() == boneIndex)
 							boneInVector = k;
 
@@ -419,7 +419,7 @@ void WSkeleton::Update(float fDeltaTime) {
 				if (curSubAnim->parentSubAnimation != -1) {
 					uint32_t boneIndex = curSubAnim->boneIndices[0];
 					uint32_t boneInVector = (uint32_t)-1;
-					for (uint32_t k = 0; k < curFrame->boneV.size() && boneInVector == -1; k++)
+					for (uint32_t k = 0; k < curFrame->boneV.size() && boneInVector == (uint32_t)-1; k++)
 						if (curFrame->boneV[k]->GetIndex() == boneIndex)
 							boneInVector = k;
 					curFrame->boneV[boneInVector]->SetParent(curFrameOldParent);
